telegram-ai-bot/requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
python-telegram-bot[webhooks]==21.3
